requirements.txt
streamlit
pymongo
python-dotenv
openai
requests
beautifulsoup4